Add forward compatibility loading code with a checkpoint conversion flag.
This flag controls whether or not to use the new checkpoint implementation instead of the current SaveableObject-saving code (which hasn't been implemented yet).

This flag will be made public after the new checkpoint implementation with the Trackable API has been added.

The flag will be removed after the following conditions are true:
- The new checkpoint implementation has been added.
- 3 week have passed since this CL has been submitted.
- Global TAP passes with the flag removed. All internal tests should pass with the new checkpoint implementation.

PiperOrigin-RevId: 462468886