Create New LegalizeTFModulePass
This new pass will be used for TF2XLA fallback by ops that need to be legalized  during a ModulePass due to the restrictions on FuncOp passes (https://mlir.llvm.org/docs/PassManagement/#operation-pass). Establishes a whitelist of Ops that will be legalized using this pass.

This CL unblocks the new ApproxTopKOp which has a dependency on CustomCallOp with a called computation.

PiperOrigin-RevId: 462512873