#tf-data-service `register_dataset` accepts a `dataset_id`.
If provided, tf.data service will use the provided ID for the dataset.
If the dataset ID already exists, no new dataset will be registered.
This is useful if multiple training jobs need to use the same dataset
for training. In this case, users should call `register_dataset` with
the same `dataset_id`.

PiperOrigin-RevId: 462690057