distributed dataset sparse tensor GPU test
This test with a dataset created with a sparse tensor is similar to previous tests that create a datasets with a ragged tensors (in ragged/ragged_factory_ops_test.py). This is a GPU only that uses mirrored_strategy.MirroredStrategy and is intended to improve test coverage for device placement.

PiperOrigin-RevId: 466155803