Refactor core/platform/stack_frame.h into TSL
Tensor Support Library (TSL) is a new component that is intended to ultimately live
outside of the TensorFlow repository. It'll contain all the platform abstraction
code, common basic utilities, and some profiler runtime components. It'll be shared
between XLA and TensorFlow in the context of OpenXLA: https://github.com/openxla

This is the first of a long series of CL refactoring code into this new directory.

PiperOrigin-RevId: 466480409