Make the HloLegalizeToLinalgPass test pass use a public type converter.
A previous patch changed the pass to use a private type converter which had
materializations needed by the newly introduced patterns, rendering the prior
public API for getting a conforming type converter unsuitable
(createHloToLinalgSignedIntegerConverter()).

This patch exposes the private type converter as a new LinalgTypeConverter and
renames the createHloToLinalgSignedIntegerConverter() to
createHloToLinalgTypeConverter() to better match the new use. This will be a
minor API break to downstreams using the existing
createHloToLinalgSignedIntegerConverter() and a simple rename should be all
that is required.

PiperOrigin-RevId: 468032316