[XLA:GPU] Don't slice+re-pad when we some output dims of a convolution must be 0.
CudnnVectorizeConvolutions will generate code that does:

- pad output features to a multiple of 32
- reshape input from [N,C,H,W] to [N,C/32,H,W,32]
- run the conv
- reshape output from [N,C/32,H,W,32] to [N,C,H,W]
- slice off the padding on the `C` channel.

If this is followed by another convolution (very common), then the slice is
immediately followed by another pad.

This slice/pad may be redundant.  We know that those trailing channels are 0.
So why slice them off and then re-pad them?

Ideally we can eliminate the whole reshape+slice+pad+reshape chain, allowing us
to take one conv's output and pipe it directly into another conv's input.  This
saves a kernel.

The reshape-slice-pad-reshape chain is particularly bad after a change to
fusion heuristics in b/241793029.  Part of the motivation for this change is
make us stop generating this HLO that's pathological to the new fusion pass.
This will allow us to land that change to fusion.

PiperOrigin-RevId: 468508652