[TFTRT]: Remove shape optimization profile WAR for new TRT versions
If an input shape value tensor has min=opt=max profiles,
then TRT 7 was pruning this tensor from the network
(since its value could be deduced from the profile).
TF-TRT introduced `FixShapeValueProfile` to modify the
profiles which prevented TRT from pruning these inputs.

This workaround is not necessary in TRT8, and actually
can be harmful, because it can lead to inconsistent input
profiles. Remove this WAR for newer versions of TRT.

Signed-off-by: Meenakshi Venkataraman <meenakshiv@nvidia.com>