Implement merge sort.
Also reorganize the sort_ops_tests so they finish in a slightly more reasonable time when mlir lowering is enabled.

The sort lowering lowers directly to buffers. Some alternatives that were discussed were to lower to some other form first (that would allow the comparator to be lowered already), and do the final lowering after bufferization. This is significantly more work, and since the chosen approach seems to work for now, I'll leave this as potential future work.

PiperOrigin-RevId: 470020259