Change to use a simpler graph partitioning method.
We don't need to support non-XLA GPU, so the workaround to wrap ops on different devices into separate functions is not needed anymore. This simplifies the graph paritioning, as well as the downstream procesing.
The new graph partitioning also adds dependencies between Send/Recv pairs to ensure TF graph to MLIR conversion will generate MLIR ops in a valid order.

PiperOrigin-RevId: 470361677