Small code cleanup in ReduceConversion.
Do not create another 'operand' array that is identical to adaptor.operands()
Also fix the type computation for the block arguments when doing the signature
conversion from the original apply function to the linalg apply function. It
worked before because we can be sure that the element type of the i-th operand
matches the element type of the i-th output, but it is confusing to read the
code.

PiperOrigin-RevId: 471748347