Implements a pass that moves initializer function ops to the main function.
This change introduces a new pass called `MergeInitializerFunctionOpsToMainPass`.
This moves all ops from the initializer function to the main function so that the ops reside on the top-level graph.
This is to avoid the initializer functions from getting removed after the canonicalizer pass when it is again loaded into the MLIR module.

PiperOrigin-RevId: 473928882