Custom Device Pinning Overriding
TF Changes:
Currently, the CustomDeviceOpHandler chooses the device to pin to based on a set of rules. One of the rules is when there are resource tensors amongst the inputs, the op is placed on the device of the first resource input.

DTensor, which uses the CustomDevice abstraction, wants to change this rule. Presumably other users of CustomDevice would also like to change this rule to their liking. This CL implements derivations of CustomDevice to implement their own set of rules for custom device pinning, when there is a single CustomDevice amongst inputs to an op.

Note that this CL does not change any existing, current behavior. It only changes custom device placement for DTensor

DTensor Changes
1. Set our own custom rule for dtensor pinning
2. Expand rule for resource tensor upcasting to dtensor. Now let any resource tensor upcast to a dtensor as long as were broadcasting to a CPU device (GPU and TPU broadcasting is not supported currently).

PiperOrigin-RevId: 474337912