Prepare tf.function to make ResourceVariable a subclass of CompositeTensor.
This CL does not intend to change the behavior of tf.function. The changes simply make sure tf.function could behave the same after we make ResourceVariable a subclass of CompositeTensor and expand ResourceVariables into dt_resource handle when expand_composites=True.

Unfortunately we need to special case ResourceVariable because
1. dt_resource tensors are not de-aliased while ResourceVariables are de-aliased in tf.function,
2. gradient related.

We could remove 2 of 3 special cases in the CL once we de-alias dt_resource tensors or add is_leaf to tf.nest. Will need to investigate a little more on how to remove the 3rd special case in _get_defun_input.

PiperOrigin-RevId: 474348028