Configure tf.nest to not expand ResourceVariables to dt_resource handles.
ResourceVariable is currently not a CompositeTensor. The main purpose of this CL is to reduce the risk of rolling back.

I am planning to first make ResourceVariable a subclass of CompositeTensor, but don't expand ResourceVariable into dt_resource handles.

Then, we will revert this change and also update util.cc to expand VariableSpec.

PiperOrigin-RevId: 474601362