DTensor - Generalize SaveOp SPMD To Work For Non-Constants
Previously, the SPMD for SaveOp assumes that the `prefix` input is a constant folded tensor. So it would extract it as a constant, read the `prefix` as a std::string, and do some computation on it to make a new prefix.. However, going forward to the new CheckpointV2, the `prefix` is not always constant folded. Such a case would be when checkpoint.save() under the hood is implemented with a tf.function(), then the `prefix` tensor is the output of a tf.ShardedFilename op. In that case, `prefix` is not a constant, and dtensor would segfault since it tried to extract it as a constant.

This CL generalizes the expansion such that it does not assume that `prefix` is a constant - it works for both constants and non-constants. New implementation uses TF ops to do the same string computation as we were previously doing before.

PiperOrigin-RevId: 474856615