Replace SameOperandsAndResultType trait with TF specific trait
The core MLIR trait no longer allows different but compatible types like tensor<*xf32> and tensor<4xf32>. We need to allow such ops in TensorFlow.

Using the existing trait TF_SameOperandsAndResultTypeResolveRef as an replacement. The op generator automatically infers result type if SameOperandsAndResultType trait is present so this also requires adding a simple shape function propagating the operand type.

The core Idempotent trait requires the SameOperandsAndResultType trait so we need TF version for that as well.

This change only replaces the trait for the AbsOp and rest of the ops will be updated in a follow-up change.

TFExecutor dialect imports TF traits so this required updates in the dialect as well. The op generator adds shape function in particular scenarios if shape function interfaces are imported.

PiperOrigin-RevId: 475381274