[xla:cpu:mlir] Add a gml_st tiling (to points) pass and hook it up to the pipeline
This means that incoming scatters will be converted to thlo and tiled to
points. It's currently not used because XLA:CPU converts all scatters to while
loops, but this lets us disable that pass in the future.

It's not ready for everything yet as the thlo scatter doesn't support sliced
scatters yet. Once it does we can remove ScatterExpander from the mlir-based
flow.

PiperOrigin-RevId: 475517897