Tighten HLO type converters
When preparing for tomorrow's OpenXLA community meeting, I realized that I forgot to account for tuple types in StableHLO <=> MHLO conversions.

At that point, I realized that I forgot to do an exhaustive analysis of which types should be converted between StableHLO and MHLO dialects. In an earlier prototype of the StableHLO => MHLO conversion which I used to bootstrap the full-fledged conversion, I added a default clause to type converters which converted a Type into a Type, which got things working, and I happily forgot about being more thorough.

Then, I decided to refactor HLO type converters to use a strict allowlist of types. I meant well, but then I realized that this really cannot work given that HLO programs can be embedded into other programs which can use all sorts of types.

I settled on a middle ground. HLO type converters still have a pass through clause, but it no longer applies to the types and encodings which are defined by the source dialect (i.e. by MHLO for HloToStablehloTypeConverter and StableHLO for StablehloToHloTypeConverter). This way we can at least rule out cases where we add a type or an encoding to one of our dialects and then forget to update the converter.

PiperOrigin-RevId: 475555695